+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# future
+- Fix #1036 : don't kill detached processes, an option has been added to keep the old behavior
+
 # 0.12.7
 
 - pm2 logs : Now shows merged logs